provider: initial DevDocs provider
This adds an initial provider which sources documentation from
https://devdocs.io/

The motivation is that often you want very specific version information
for a documentation set when interacting with AI, rather than what the
AI was trained on. Additionally by providing exact documentation it can
improve the quality of responses.

DevDocs is a great free resource of documentation in a regular format.
This makes it easier for us to consume and search.

This provider will cache the index for a documentation set for 12 hours
for performance reasons (and friendliness to devdocs hosting). We use
fuse.js to get some scoring on results for free. I looked into using the
same scorer as devdocs, but the code is not easily reusable.

Test Plan: A unit test based on real fixture data is added. Additionally
if you set `INTEGRATION=1` it will run against the real devdocs.io
source.