Add mention selectors to OpenCtx provider protocol (#103)
This PR modifies the `MetaResult` to the following:

- ~`mentions.kind` is the list of types of mention items the provider
can return. For example github pr, issues, code search results.~
- `mentions.selectors` is the list of regex patterns, which if matches
with the mention text would mean that the provider can return items for
that. We will use it to get context items from all the supported
providers for when the user pasts in a URL after @.
- Whey we do not have `features.mentions: boolean | {}` ? because then
you can not do `meta?.features?.mentions?.selector`. We will always have
to check of the mentions type first.
- Accordingly the `meta.selector` is moved to
`meta.features.annotations.selectors`.

``` typescript
export interface MetaResult {
    /**
     * The name of the provider.
     */
    name: string

        /**
         * Configuration for the mentions feature.
         */
        mentions?: {
            /**
             * A list of patterns matching the mention text for which the provider can return mentions
             */
            selectors?: MentionSelector[]
        }
        /**
         * Configuration for the annotations feature.
         */
        annotations?: {
            /**
             * A list of patterns matching the mention text for which the provider can return mentions
             */
            selectors?: AnnotationSelector[]
        }
}
```