provider: better search and default settings for devdocs (#106)
Now if you do not specify `urls` in your settings we will use the same
defaults as DevDocs does.

Additionally the search was too fuzzy before. This lead to confusing
behaviour where it looked like the integration was broken. Instead we
switch to a more classic fuzzy search library.

Test Plan: tests where added to ensure fuzzy finder can find no results
and that we work with the default docset. Additionally we assert that
the top result on a known query.