bin: introduce mention-items subcommand
It is quite clunky to copy paste the output mention JSON into the items
subcommand. So this introduces a new subcommand "mention-items" which
automates that process. It will take an item from the mentions results
and send it to items. By default it is the first mention, but can does
accept an argument for another mention.

Test Plan: Ran locally. See this example:

  $ OPENCTX_CONFIG=../provider/devdocs/dist/bundle.js pnpm openctx mentions 'abort controller'
  $ OPENCTX_CONFIG=../provider/devdocs/dist/bundle.js pnpm openctx mention-items 'abort controller'