client: add ObserveOptions.errorHook
This introduces a new option when calling an observable method on the
client. It allows a caller to collect errors on the underlying
providers.

Currently the client implementation when encountering an error in a
provider will log it and then swallow it unbeknownst to the client. This
is reasonable behaviour since we don't want one bad provider causing all
of OpenCtx to fail. For example if the methods where to throw errors you
wouldn't be able to collect the working provider results.

As an alternative I considered updating the return type of the client to
either return the underlying type or an error. This ended up being quite
a large change and didn't feel like it justified it for the extra burden
it imposed on each call site. However, I can go that route so that each
client explicitly handles errors.

Test Plan: the unit test was updated to ensure errorHook is not called
for the happy path, and is called for the erroring provider.
Additionally because we now no longer call console.error the unit test
output is much cleaner when all tests pass.