client: introduce ProviderMethodOptions
My last PR introducing errorHook did not expose it in the vscode
controller. This commit does the larger change of making it possible to
be set by callers. Currently all methods optionally take a providerUri
which affects the behaviour. This is now replaced with the
ProviderMethodOptions argument which contains the providerUri and the
newly introduced errorHook.

The name choice of the word "Method" was used since that is what we call
the underlying provider methods in our documentation.

Test Plan: CI will cover if we broke anything. Additionally ran the cli
with mention-items since that tests providerUri being set. Surpisingly
none of our tests seem to test providerUri being set, but our CI will
cover the normal case of no options being set working.

  OPENCTX_CONFIG=../provider/devdocs/dist/bundle.js pnpm openctx mention-items 'AbortController'