fix(pprof): avoid breaking search loop prematurely
When searching for the Go binary, we assume that it will be in the same directory as the report, or in one of its child directories.
If we break the loop the moment we find the report, then we might not find the binary even if it's in the same directory but is alphabetically sorted after the report file.