vscode-lib: use errorHook to report errors (#142)
This commit factors out the error reporting logic inside of
createController into an errorReporter interface. ErrorReporter also
provides functionality so we can use the errorHooks.

From my testing, it is useful to do error reporting per provider. For
example a misconfigured provider will trip the errorWaiter, leading to
errors when calling mentions not being reported.

Test Plan: linked openctx into cody and tested with my usual providers,
a missing provider and a provider which throws an error mentions. I
observed on cody chat startup complaints about the missing provider and
complaints again when the error provider is called for mentions.