client: introduce preloadDelay
This setting allows us to preload the openctx providers after a
delay (which could be zero milliseconds). The intention is for Cody to
set a reasonable value here (5s?) so that when a user does the first @
mention the providers have already been fetched.

Test Plan: added logging and observed the preload completing before
activating the cody chat panel. Then loaded the cody chat panel and
instantly do a mention. All OpenCtx providers where immediate.