linear-issues: publish d.ts and correctly specify devDeps
We had a bunch of deps which are only dev deps. Additionally because we
called tsc --clean when bundling we didn't publish the index.d.ts file
which confused vscode when I was importing the provider into Cody.

Finally our cody build is stricter on how we reference built in node
modules. Just for fs/promises I updated all call sites to use the node:
prefix.