expose async-generator client API
Previously, the client and controller only exposed a "watch" API using RxJS Observables so that callers could subscribe to and watch changes (to mentions, annotations, etc.). This required callers to depend on RxJS, which is a big dependency that can introduce complexity in their own applications. Now, there is a new API `xyzChanges__asyncGenerator` for each of `meta`, `mentions`, `items`, and `annotations` that lets callers just use native async generators. This API is experimental and may be changed without notice.